+++ src/lib/utils.js
@@ -1,12 +1,9 @@
 import path from "path";
 import fs from "fs";
 import denodeify from "denodeify";
-import fetch from "node-fetch";
 
 fs.readFile = denodeify(fs.readFile);
 fs.readdir = denodeify(fs.readdir);
-
-const api = "https://modwatchapi-ansballard.rhcloud.com"
 
 export function cleanModFile(opts = {}) {
   if(!opts.filepath) {
@@ -14,7 +11,7 @@ export function cleanModFile(opts = {}) {
   }
   return fs.readFile(path.resolve(opts.filepath), { encoding: "utf8" })
   .then(content => {
-    let filename = path.basename(opts.filepath);
+    const filename = path.basename(opts.filepath);
     return {
       content: content
         .split("\r\n")
@@ -43,60 +40,6 @@ export function scanModDirectory(opts = {}) {
       game: opts.game
     });
   }
-}
-
-export function getUsers(opts = {}) {
-  opts.from = opts.from || 0;
-  return fetch(`${api}/api/users/list`)
-  .then(response => response.json())
-  .then(users => users.slice(opts.from || 0, opts.limit || opts.from + 10))
-  .then(users => users.map(user => user.username));
-}
-
-export function getUserFiles(opts = {}) {
-  if(!opts.username) {
-    return Promise.reject("Username required");
-  }
-  return fetch(`${api}/api/user/${opts.username}/files`)
-  .then(response => response.json())
-}
-
-export function getUserFile(opts = {}) {
-  if(!opts.username || !opts.file) {
-    return Promise.reject("Username/File type required");
-  }
-  return fetch(`${api}/api/user/${opts.username}/file/${opts.file}`)
-  .then(response => response.json())
-}
-
-export function getUserProfile(opts = {}) {
-  if(!opts.username) {
-    return Promise.reject("Username required");
-  }
-  return fetch(`${api}/api/user/${opts.username}/profile`)
-  .then(response => response.json())
-}
-
-export function uploadMods(opts = {}) {
-  if(!opts.username || !opts.password) {
-    return Promise.reject("Username/Password required");
-  }
-  return fetch(`${api}/loadorder`, {
-    method: "POST",
-    headers: {
-      "Accept": "application/json",
-      "Content-Type": "application/json"
-    },
-    body: JSON.stringify({
-      username: opts.username,
-      password: opts.password,
-      plugins: opts.plugins,
-      modlist: opts.modlist,
-      ini: opts.ini,
-      prefsini: opts.prefsini,
-      game: opts.game
-    })
-  });
 }
 
 function scanDefaultModDirectories(opts = {}) {