+++ tests/utilsSpec.js
@@ -1,2 +1,0 @@
-import test from "ava";
-import { getUsers, getUserFiles, getUserFile, getUserProfile } from "../../src/lib/utils";